Sorting by distance to user is implemented. Next priority is testing.

+++ app/src/main/java/com/example/srikar/parklocations/model/ParkLocationResponse.java
@@ -8,18 +8,18 @@ import com.google.gson.annotations.SerializedName;
  */
 public class ParkLocationResponse {
     @SerializedName("latitude")
-    private Float latitude;
+    private Double latitude;
     @SerializedName("human_address")
     private String unformattedAddress;
     //ignore "needs_recording"
     @SerializedName("longitude")
-    private Float longitude;
+    private Double longitude;
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
@@ -31,11 +31,11 @@ public class ParkLocationResponse {
         this.unformattedAddress = unformattedAddress;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
